+++ src/collect.nim
@@ -108,8 +108,8 @@ proc collectChalkTimeHostInfo*() =
             plugin.isSystem():
           hostInfo[k] = v
     except:
-      warn("When collecting chalk-time host info, plugin implementation " &
-           plugin.name & " threw an exception it didn't handle: " & getCurrentExceptionMsg())
+      error("When collecting chalk-time host info, plugin implementation " &
+            plugin.name & " threw an exception it didn't handle: " & getCurrentExceptionMsg())
       dumpExOnDebug()
 
 proc initCollection*() =
@@ -166,9 +166,9 @@ proc collectRunTimeArtifactInfo*(artifact: ChalkObj) =
           artifact.collectedData[k] = v
       trace(plugin.name & ": Plugin called.")
     except:
-      warn("When collecting run-time artifact data, plugin implementation " &
-           plugin.name & " threw an exception it didn't handle (artifact = " &
-           artifact.name & "): " & getCurrentExceptionMsg())
+      error("When collecting run-time artifact data, plugin implementation " &
+            plugin.name & " threw an exception it didn't handle (artifact = " &
+            artifact.name & "): " & getCurrentExceptionMsg())
       dumpExOnDebug()
 
   let hashOpt = artifact.callGetEndingHash()
@@ -224,9 +224,9 @@ proc collectChalkTimeArtifactInfo*(obj: ChalkObj, override = false) =
           obj.collectedData[k] = v
       trace(plugin.name & ": Plugin called.")
     except:
-      warn("When collecting chalk-time artifact data, plugin implementation " &
-           plugin.name & " threw an exception it didn't handle (artifact = " &
-           obj.name & "): " & getCurrentExceptionMsg())
+      error("When collecting chalk-time artifact data, plugin implementation " &
+            plugin.name & " threw an exception it didn't handle (artifact = " &
+            obj.name & "): " & getCurrentExceptionMsg())
       dumpExOnDebug()
 
 proc collectRunTimeHostInfo*() =
@@ -251,9 +251,9 @@ proc collectRunTimeHostInfo*() =
             plugin.isSystem():
           hostInfo[k] = v
     except:
-      warn("When collecting run-time host info, plugin implementation " &
-           plugin.name & " threw an exception it didn't handle: " &
-           getCurrentExceptionMsg())
+      error("When collecting run-time host info, plugin implementation " &
+            plugin.name & " threw an exception it didn't handle: " &
+            getCurrentExceptionMsg())
       dumpExOnDebug()