+++ CHANGELOG.md
@@ -7,6 +7,9 @@
 - `docker` run-time host metadata collection was failing
   for non-build commands such as `docker push`.
   ([#399](https://github.com/crashappsec/chalk/pull/399))
+- `procfs` plugin was throwing an exception while parsing
+  `/proc/net/dev` to populate `_OP_IPV[4/6]_INTERFACES` keys.
+  ([#399](https://github.com/crashappsec/chalk/pull/399))
 
 ## 0.4.10